requirements.txt
livekit-agents[openai,deepgram,silero,ollama]
python-dotenv
httpx
flask
